Creates scripts folder and assembly definition
Creates a dedicated "Scripts" folder to organize the project's scripts.
Adds an assembly definition file to improve code organization and compilation times. Changing code will prompt a recompilation only of scripts of the assembly.
The assembly definition needs to explicitly reference other assemblies it uses. In this case, the Unity input system needs to be referenced.
Assembly definitions also offer a clean possibility to define editor- and test scripts that are only used in the editor and are not bundled with the executable.